Don't unlink non-existent instruction files - fixes issue #24
git-svn-id: https://muddle.googlecode.com/svn/trunk/muddle@156 837abeb0-121e-11de-ab22-55a8a853a75f

+++ muddled/db.py
@@ -77,7 +77,8 @@ class Database(object):
         file_name = self.instruction_file_name(label)
 
         if instr_file is None:
-            os.unlink(file_name)
+            if os.path.exists(file_name):
+                os.unlink(file_name)
         else:
             instr_file.save_as(file_name)
